refactor for testing
